refactor(expense-documents): extract ExpenseDocumentCreateService (create family + update, verbatim)
Phase 3 (final) of transactional-core decompose. create, createDraftForRepair,
createCreditNote, createPayroll, createSettlement, createPettyCash, update moved
verbatim into ExpenseDocumentCreateService; facade delegates + sheds docNumber/
aggregator/transition/ssoConfig/pettyCash/payrollCustom + the readBoolFlag wrapper.
Facade is now a thin delegating shell (onModuleInit + ctor + delegations).
+characterization for createPettyCash/update happy-paths + createDraftForRepair
(external-tx). 436 existing + new green, tsc 0, lint 0.
