+++ apps/api/src/modules/expense-documents/expense-documents.module.ts
@@ -17,6 +17,7 @@ import { LineAggregatorService } from './services/line-aggregator.service';
 import { JePreviewService } from './services/je-preview.service';
 import { ExpenseDocumentQueryService } from './services/expense-document-query.service';
 import { ExpenseDocumentLifecycleService } from './services/expense-document-lifecycle.service';
+import { ExpenseDocumentCreateService } from './services/expense-document-create.service';
 import { ExpenseVoucherPdfService } from './services/expense-voucher-pdf.service';
 import { PettyCashService } from './services/petty-cash.service';
 import { PayrollCustomService } from './services/payroll-custom.service';
@@ -53,6 +54,7 @@ import { ReversePermissionGuard } from './reverse-permission.guard';
     JePreviewService,
     ExpenseDocumentQueryService,
     ExpenseDocumentLifecycleService,
+    ExpenseDocumentCreateService,
     ExpenseVoucherPdfService,
     PettyCashService,
     PayrollCustomService,
